requirements.txt

pandas
numpy
SQLAlchemy
psycopg2-binary
python-dotenv

matplotlib
seaborn

scikit-learn
scipy
nltk
regex

# Training (used in notebooks / optional jobs)
joblib

# API serving (used by main.py)
fastapi
pydantic
uvicorn[standard]

mlflow

celery
redis